feature: Add files to group classes
- Removed embed from callWebService methods
