+++ Classes/Encounters.swift
@@ -1,0 +1,59 @@
+//
+//  Encounters.swift
+//  PokemonKit
+//
+//  Created by Christopher Jennewein on 6/27/18.
+//
+
+import Foundation
+
+
+/// Methods by which the player might can encounter Pokémon in the wild, e.g., walking in tall grass. Check out Bulbapedia for greater detail.
+open class PKMEncounterMethod: Codable {
+    
+    /// The identifier for this encounter method resource
+    open var id: Int?
+    
+    /// The name for this encounter method resource
+    open var name: String?
+    
+    /// A good value for sorting
+    open var order: Int?
+    
+    /// The name of this encounter method listed in different languages
+    open var names: [PKMName]?
+}
+
+
+/// Encounter Condition
+open class PKMEncounterCondition: Codable {
+    
+    /// The identifier for this encounter condition resource
+    open var id: Int?
+    
+    /// The name for this encounter condition resource
+    open var name: String?
+    
+    /// A list of possible values for this encounter condition
+    open var values: [PKMNamedAPIResource]?
+    
+    /// The name of this encounter method listed in different languages
+    open var names: [PKMName]?
+}
+
+
+/// Encounter Condition Value
+open class PKMEncounterConditionValue: Codable {
+    
+    /// The identifier for this encounter condition value resource
+    open var id: Int?
+    
+    /// The name for this encounter condition value resource
+    open var name: String?
+    
+    /// The condition this encounter condition value pertains to
+    open var condition: [PKMNamedAPIResource]?
+    
+    /// The name of this encounter condition value listed in different languages
+    open var names: [PKMName]?
+}
