+++ Classes/Items.swift
@@ -1,0 +1,8 @@
+//
+//  Items.swift
+//  PokemonKit
+//
+//  Created by Christopher Jennewein on 6/27/18.
+//
+
+import Foundation
