+++ Classes/PokemonKit.swift
@@ -1548,32 +1548,7 @@ open class PKMAbilityEffectChange: Codable, SelfDecodable {
 }
 
 
-/// Contest Combo Detail
-open class PKMContestComboDetail: Codable, SelfDecodable {
-    
-    /// A list of moves to use before this move
-    open var useBefore: [PKMNamedAPIResource]?
-    
-    /// A list of moves to use after this move
-    open var useAfter: [PKMNamedAPIResource]?
-    
-    public static var decoder: JSONDecoder = {
-        let decoder = JSONDecoder()
-        decoder.keyDecodingStrategy = .convertFromSnakeCase
-        return decoder
-    }()
-}
 
-
-/// Contest Combo Sets
-open class PKMContestComboSets: Codable {
-    
-    /// A detail of moves this move can be used before or after, granting additional appeal points in contests
-    open var normal: PKMContestComboDetail?
-    
-    /// A detail of moves this move can be used before or after, granting additional appeal points in super contests
-    open var `super`: PKMContestComboDetail?
-}
 
 
 /// Moves are the skills of Pokémon in battle. In battle, a Pokémon uses one move each turn. Some moves (including those learned by Hidden Machine) can be used outside of battle as well, usually for the purpose of removing obstacles or exploring new areas.
@@ -2101,209 +2076,12 @@ open class PKMGeneration: Codable, SelfDecodable {
 }
 
 
-/// Evolution triggers are the events and conditions that cause a pokemon to evolve. Check out Bulbapedia for greater detail.
-open class PKMEvolutionTrigger: Codable, SelfDecodable {
-    
-    /// The identifier for this evolution trigger resource
-    open var id: Int?
-    
-    /// The name for this evolution trigger resource
-    open var name: String?
-    
-    /// The name of this evolution trigger listed in different languages
-    open var names: [PKMName]?
-    
-    /// A list of pokemon species that result from this evolution trigger
-    open var pokemonSpecies: PKMNamedAPIResource?
-    
-    public static var decoder: JSONDecoder = {
-        let decoder = JSONDecoder()
-        decoder.keyDecodingStrategy = .convertFromSnakeCase
-        return decoder
-    }()
-}
 
 
-/// Evolution Detail
-open class PKMEvolutionDetail: Codable, SelfDecodable {
-    
-    /// The item required to cause evolution this into Pokémon species
-    open var item: PKMNamedAPIResource?
-    
-    /// The type of event that triggers evolution into this Pokémon species
-    open var trigger: PKMNamedAPIResource?
-    
-    /// The gender the evolving Pokémon species must be in order to evolve into this Pokémon species
-    open var gender: PKMNamedAPIResource?
-    
-    /// The item the evolving Pokémon species must be holding during the evolution
-    open var heldItem: PKMNamedAPIResource?
-    
-    /// The move that must be known by the evolving Pokémon species during the evolution trigger event in order to evolve into this Pokémon species
-    open var knownMove: PKMNamedAPIResource?
-    
-    /// The evolving Pokémon species must know a move with this type during the evolution trigger event in order to evolve into this Pokémon species
-    open var knownMoveType: PKMNamedAPIResource?
-    
-    /// The location the evolution must be triggered at.
-    open var location: PKMNamedAPIResource?
-    
-    /// The minimum required level of the evolving Pokémon species to evolve into this Pokémon species
-    open var minLevel: Int?
-    
-    /// The minimum required level of happiness the evolving Pokémon species to evolve into this Pokémon species
-    open var minHappiness: Int?
-    
-    /// The minimum required level of beauty the evolving Pokémon species to evolve into this Pokémon species
-    open var minBeauty: Int?
-    
-    /// The minimum required level of affection the evolving Pokémon species to evolve into this Pokémon species
-    open var minAffection: Int?
-    
-    /// Whether or not it must be raining in the overworld to cause evolution this Pokémon species
-    open var needsOverworldRain: Bool?
-    
-    /// The pokemon species that must be in the players party in order for the evolving Pokémon species to evolve into this Pokémon species
-    open var partySpecies: PKMNamedAPIResource?
-    
-    /// The player must have a pokemon of this type in their party during the evolution trigger event in order for the evolving Pokémon species to evolve into this Pokémon species
-    open var partyType: PKMNamedAPIResource?
-    
-    /// The required relation between the Pokémon's Attack and Defense stats. 1 means Attack > Defense. 0 means Attack = Defense. -1 means Attack < Defense.
-    open var relativePhysicalStats: Int?
-    
-    /// The required time of day. Day or night.
-    open var timeOfDay: String?
-    
-    /// Pokémon species for which this one must be traded.
-    open var tradeSpecies: PKMNamedAPIResource?
-    
-    /// Whether or not the 3DS needs to be turned upside-down as this Pokémon levels up.
-    open var turnUpsideDown: Bool?
-    
-    public static var decoder: JSONDecoder = {
-        let decoder = JSONDecoder()
-        decoder.keyDecodingStrategy = .convertFromSnakeCase
-        return decoder
-    }()
-}
 
 
-/// Clain Link
-open class PKMClainLink: Codable, SelfDecodable {
-    
-    /// Whether or not this link is for a baby Pokémon. This would only ever be true on the base link.
-    open var isBaby: Bool?
-    
-    /// The Pokémon species at this point in the evolution chain
-    open var species: PKMNamedAPIResource?
-    
-    /// All details regarding the specific details of the referenced Pokémon species evolution
-    open var evolutionDetails: [PKMEvolutionDetail]?
-    
-    /// A List of chain objects.
-    open var evolvesTo: [PKMClainLink]?
-    
-    public static var decoder: JSONDecoder = {
-        let decoder = JSONDecoder()
-        decoder.keyDecodingStrategy = .convertFromSnakeCase
-        return decoder
-    }()
-}
 
 
-/// Evolution Chain
-open class PKMEvolutionChain: Codable, SelfDecodable {
-    
-    /// The identifier for this evolution chain resource
-    open var id: Int?
-    
-    /// The item that a Pokémon would be holding when mating that would trigger the egg hatching a baby Pokémon rather than a basic Pokémon
-    open var babyTriggerItem: PKMNamedAPIResource?
-    
-    /// The base chain link object. Each link contains evolution details for a Pokémon in the chain. Each link references the next Pokémon in the natural evolution order.
-    open var chain: PKMClainLink?
-    
-    public static var decoder: JSONDecoder = {
-        let decoder = JSONDecoder()
-        decoder.keyDecodingStrategy = .convertFromSnakeCase
-        return decoder
-    }()
-}
-
-
-/// Encounter Condition Value
-open class PKMEncounterConditionValue: Codable {
-    
-    /// The identifier for this encounter condition value resource
-    open var id: Int?
-    
-    /// The name for this encounter condition value resource
-    open var name: String?
-    
-    /// The condition this encounter condition value pertains to
-    open var condition: [PKMNamedAPIResource]?
-    
-    /// The name of this encounter condition value listed in different languages
-    open var names: [PKMName]?
-}
-
-
-/// Encounter Condition
-open class PKMEncounterCondition: Codable {
-    
-    /// The identifier for this encounter condition resource
-    open var id: Int?
-    
-    /// The name for this encounter condition resource
-    open var name: String?
-    
-    /// A list of possible values for this encounter condition
-    open var values: [PKMNamedAPIResource]?
-    
-    /// The name of this encounter method listed in different languages
-    open var names: [PKMName]?
-}
-
-
-/// Methods by which the player might can encounter Pokémon in the wild, e.g., walking in tall grass. Check out Bulbapedia for greater detail.
-open class PKMEncounterMethod: Codable {
-    
-    /// The identifier for this encounter method resource
-    open var id: Int?
-    
-    /// The name for this encounter method resource
-    open var name: String?
-    
-    /// A good value for sorting
-    open var order: Int?
-    
-    /// The name of this encounter method listed in different languages
-    open var names: [PKMName]?
-}
-
-
-/// Super contest effects refer to the effects of moves when used in super contests.
-open class PKMSuperContestEffect: Codable, SelfDecodable {
-    
-    /// The identifier for this super contest effect resource
-    open var id: Int?
-    
-    /// The level of appeal this super contest effect has
-    open var appeal: String?
-    
-    /// The flavor text of this super contest effect listed in different languages
-    open var flavorTextEntries: [PKMFlavorText]?
-    
-    /// A list of moves that have the effect when used in super contests
-    open var moves: [PKMNamedAPIResource]?
-    
-    public static var decoder: JSONDecoder = {
-        let decoder = JSONDecoder()
-        decoder.keyDecodingStrategy = .convertFromSnakeCase
-        return decoder
-    }()
-}
 
 
 /// Flavor Text
@@ -2323,64 +2101,6 @@ open class PKMFlavorText: Codable, SelfDecodable {
 }
 
 
-/// Effect Entry
-open class PKMEffectEntry: Codable {
-    
-    /// The localized effect text for an API resource in a specific language
-    open var effect: String?
-    
-    /// The language this effect is in
-    open var language: PKMName?
-}
-
-
-/// Contest effects refer to the effects of moves when used in contests.
-open class PKMContestEffect: Codable, SelfDecodable {
-    
-    /// The identifier for this contest type resource
-    open var id: Int?
-    
-    /// The base number of hearts the user of this move gets
-    open var appeal: String?
-    
-    /// The base number of hearts the user's opponent loses
-    open var jam: Int?
-    
-    /// The result of this contest effect listed in different languages
-    open var effectEntries: [PKMEffectEntry]?
-    
-    /// The flavor text of this contest effect listed in different languages
-    open var flavorTextEntries: [PKMFlavorText]?
-    
-    public static var decoder: JSONDecoder = {
-        let decoder = JSONDecoder()
-        decoder.keyDecodingStrategy = .convertFromSnakeCase
-        return decoder
-    }()
-}
-
-
-/// Contest types are categories judges used to weigh a Pokémon's condition in Pokémon contests. Check out Bulbapedia for greater detail.
-open class PKMContestType: Codable, SelfDecodable {
-    
-    /// The identifier for this contest type resource
-    open var id: Int?
-    
-    /// The name for this contest type resource
-    open var name: String?
-    
-    /// The berry flavor that correlates with this contest type
-    open var berryFlavor: PKMNamedAPIResource?
-    
-    /// The name of this contest type listed in different languages
-    open var names: [PKMName]?
-    
-    public static var decoder: JSONDecoder = {
-        let decoder = JSONDecoder()
-        decoder.keyDecodingStrategy = .convertFromSnakeCase
-        return decoder
-    }()
-}
 
 
 /// Paged Object
